backend/pyproject.toml
[project]
name = "datasaur-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.40.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "torch>=2.3.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
